# Support for JSON Numbers