Adds support for #17.

This probably needs significant edits prior to merging, but the benchmarks with the primitive tokenizing (uses none of the TryReadU64 optimizations) still are pretty good:

```text
$ /home/ahuszagh/git/fast-float-rust/target/release/fast-float-simple-bench file ext/data/canada.txt 
=====================================================================================
|                         canada.txt (111126, 1.93 MB, f64)                         |
|===================================================================================|
|                                                                                   |
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            25.84    26.30    27.14    27.64    28.17    29.85    40.15 |
| fast-float-tokenized    44.33    45.83    46.25    46.71    47.36    49.06    88.16 |
| lexical               74.67    76.35    76.96    77.64    78.73    81.79   132.29 |
| from_str             218.95   220.82   222.62   224.34   226.81   237.21   407.54 |
|                                                                                   |
| Mfloat/s                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float            24.90    33.59    35.51    36.18    36.85    38.02    38.69 |
| fast-float-tokenized    11.34    20.38    21.11    21.41    21.62    21.82    22.56 |
| lexical                7.56    12.23    12.70    12.88    12.99    13.10    13.39 |
| from_str               2.45     4.22     4.41     4.46     4.49     4.53     4.57 |
|                                                                                   |
| MB/s                    min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| fast-float           433.38   584.58   618.00   629.61   641.32   661.66   673.33 |
| fast-float-tokenized   197.39   354.72   367.43   372.58   376.30   379.67   392.55 |
| lexical              131.54   212.84   221.04   224.14   226.11   227.91   233.05 |
| from_str              42.70    73.39    76.73    77.57    78.17    78.81    79.47 |
|                                                                                   |
=====================================================================================
```

I've added tests to ensure the differences in parsing don't lead to correctness issues, as seen [here](https://github.com/Alexhuszagh/fast-float-rust/blob/json/tests/test_json.rs).

The following methods have been added:

```rust
pub trait FastFloat {
    fn parse_tokenized<S: AsRef<[u8]>>(integral: S, fractional: S, exponent: i64, negative: bool) -> Self;
}

pub fn parse_tokenized<T: FastFloat, S: AsRef<[u8]>>(integral: S, fractional: S, exponent: i64, negative: bool) -> T;
```

In order to share more code, `parse_long_mantissa` has been changed to accept a `Decimal` by-value, and the two internal functions have been added (I probably need to rename them to be consistent, whoops):

```rust
pub fn parse_number_tokenized(i: &[u8], f: &[u8], e: i64, negative: bool) -> Option<Number>;
pub fn parse_tokenized_decimal(mut i: &[u8], mut f: &[u8], e: i64, negative: bool) -> Decimal;
```

Any feedback would be great, I'm probably going to need to refactor a bit to increase code re-use.