@Alexhuszagh 2021-05-15
I've updated the PR to do the following:
- Rename `parse_tokenized` to `parse_from_parts`
- Significantly increase the amount of code sharing, rather than duplication, for simpler maintenance.

This should be a much more satisfactory PR right now. The amount of additional code is much smaller, and mostly in unavoidable places.