@aldanor 2021-05-16
Hi @Alexhuszagh - thanks for taking time in getting a stab at this. I think the idea is good (and surely +1 for the pre-RFC - let's see how it will be received...).

Before we get into code structure discussions etc though, I've tried running the bench suite a few times for this branch and master, results seems to be fairly consistent. In particular for mesh.txt and canada.txt (the latter probably the most representative of what the most 'common' floats to be parsed would look like), there seems to be a ~7% performance drop which is outside of noise bounds.

There might be bias because the master branch code is heavily hand-tuned towards high performance on these benchmarks (on my machine in particular), but, on the other hand, medians outside of IQR signify that it's not just noise. Would you perhaps care to run these a few times on your machine to see how the numbers look like? If it's significant indeed, we should probably investigate what it's caused by? Because the code is logically identical, but the order of things may be slightly changed.

master branch:

```
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| mesh.txt              10.41    10.43    10.49    10.84    11.67    12.89    16.63 |
| canada.txt            21.77    21.83    22.57    23.41    24.41    26.21    33.01 |
| uniform               19.56    19.82    20.93    22.75    26.15    31.25    43.28 |
| one_over_rand32       29.24    29.32    30.64    31.83    33.54    37.35    56.22 |
| simple_uniform32      19.53    19.58    20.03    20.82    22.16    24.80    45.92 |
| simple_int32          15.89    15.92    16.20    17.20    18.22    19.93    31.27 |
| int_e_int             27.28    27.38    28.40    29.80    31.39    36.58    51.47 |
| simple_int64          25.68    25.73    26.61    27.62    29.14    31.91    52.64 |
| bigint_int_dot_int    40.76    40.86    41.89    43.28    45.07    48.16    58.14 |
| big_ints              83.95    85.00    87.83    90.10    92.70    98.71   117.15 |
```

this branch:

```
| ns/float                min       5%      25%   median      75%      95%      max |
|-----------------------------------------------------------------------------------|
| mesh.txt              11.07    11.15    11.25    11.62    12.34    13.67    22.26 |
| canada.txt            23.44    23.52    24.33    25.13    26.17    28.18    37.90 |
| uniform               19.37    19.55    20.99    22.96    25.57    31.03    47.99 |
| one_over_rand32       29.42    29.82    30.91    32.09    33.78    41.18    72.31 |
| simple_uniform32      19.37    19.52    20.03    21.06    22.46    26.23    69.84 |
| simple_int32          16.28    16.34    17.03    18.12    19.52    25.08    35.35 |
| int_e_int             27.17    27.33    28.12    29.35    31.00    36.11    64.15 |
| simple_int64          25.92    26.01    27.11    28.37    30.25    35.24    59.69 |
| bigint_int_dot_int    40.76    41.21    42.49    44.14    46.23    51.62    75.35 |
| big_ints              81.36    83.12    86.57    89.38    92.65   102.98   138.79 |
```