@Alexhuszagh 2021-05-16
@aldanor Definitely. This might also be due to bad inlining. I'll try changing a few inlines to `#[inline(always)]`, to ensure they're identical to the code they're replacing, since 7% sounds well outside of noise benchmarks.