@Alexhuszagh 2021-05-19
Ok this is definitely slower and inling doesn't seem to work. Getting repeated ~7% performance hits with `canada.txt`, so I'll try to triage now.