@aldanor 2021-05-19
Thanks for testing out :)

Yea, I've spent lots and lots of time (the majority other that correctness) in this crate trying to speed this up by making inlining right and avoid every single bound check where possible (hence all the weird types in `common` module), so reshuffling things without profiling will likely slow it down.

But I *think* it should be doable since it's essentially the same logic.